requirements.txt
absl-py==0.14.0
argon2-cffi==21.1.0
astor==0.8.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cloudpickle==2.0.0
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cycler==0.10.0
Cython==0.29.10
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.1.0
ffmpeg==1.4
ffmpeg-python==0.1.17
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.41.0
h5py==2.10.0
idna==3.2
imageio==2.3.0
imageio-ffmpeg==0.2.0
importlib-metadata==4.8.1
imutils==0.5.3
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.5
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.0.3
mistune==0.8.4
nbclient==0.5.4
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.4
numpy==1.18.1
opencv-python==4.5.3.56
openpyxl @ file:///tmp/build/80754af9/openpyxl_1632777717936/work
opt-einsum==3.3.0
packaging==21.0
pandas==0.24.1
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
scikit-image==0.14.3
scikit-learn==0.19.2
scikit-video==1.1.11
scipy==1.5.4
Send2Trash==1.8.0
six==1.16.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
tornado==6.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==4.3.3
typing-extensions==3.10.0.2
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yarg==0.1.9
zipp==3.6.0
